create two separate controller classes and update unit tests
